requirements.txt
gTTS
discord.py
PyNaCl